pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "primecodex"
version = "0.2.0-beta"
description = "PrimeCodex EMOTA Unity experimental framework"
readme = "README.md"
requires-python = ">=3.10"
license = "Proprietary"
authors = [{ name = "Genesis Prime" }]
dependencies = [
  "openai>=1.0.0,<2.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "PyYAML>=6.0.0,<7.0.0",
  "jsonschema>=4.0.0,<5.0.0",
  "typer>=0.12.0,<1.0.0"
]

[tool.setuptools.packages.find]
include = ["emota*"]
exclude = ["schema*"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pyright>=1.1.350",
  "ruff>=0.1.0",
  "bandit[toml]>=1.7.0",
  "safety>=2.0.0"
]

[project.scripts]
primecodex = "primecodex_cli:main"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "S",   # bandit (security)
]
ignore = [
  "S101",  # Use of assert (acceptable in tests)
  "S608",  # Possible SQL injection (not applicable here)
]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["S101"]  # Allow assert in tests
"**/test_*.py" = ["S101"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # Skip assert_used test
